added check for out param and removed aminmax file
