+++ src/frontends/pytorch/src/op/min_max.cpp
@@ -207,6 +207,26 @@ OutputVector translate_amax(const NodeContext& context) {
     return {res};
 }
 
+OutputVector translate_aminmax(const NodeContext& context) {
+    num_inputs_check(context, 1, 4);  // Expect between 1 and 4 inputs
+                                      // (input tensor, dim = none, keepdim = false, out = none)
+
+    OPENVINO_ASSERT(!context.input_is_none(3), "out argument is not supported in this conversion");
+
+    auto input = context.get_input(0);
+
+    // check if dim is provided, if not, get the range of axes to compute min and max
+    auto dim = !context.input_is_none(1) ? context.get_input(1) : get_axes_range(context, 0);
+
+    // check if keepdim is provided, if not, set it to false like PyTorch
+    bool keep_dims = !context.input_is_none(2) ? context.const_input<bool>(2) : false;
+
+    auto amin = context.mark_node(std::make_shared<v1::ReduceMin>(input, dim, keep_dims));
+    auto amax = context.mark_node(std::make_shared<v1::ReduceMax>(input, dim, keep_dims));
+
+    return {amin, amax};
+}
+
 }  // namespace op
 }  // namespace pytorch
 }  // namespace frontend
